Generating Spring Boot project structure + parser fixes
Updated the generator with creating the Spring Boot project structure. Created generators and freemarker templates for application.yml files and the Application classes with the main method.
Fixed parser to work with navigability of linked properties in associations. Other minor improvements in the parser. Updated the example model - deleted the ghost entity.